fix: resolve build errors in capacity-modeler and stratoswarm-compliance
capacity-modeler:
- Fix type mismatches: cast u8/usize to u32 for forecast message fields

stratoswarm-compliance:
- Fix DataCategory enum variants: use PII/PHI/Financial instead of
  PersonallyIdentifiable/HealthRecords/FinancialRecords
- Fix ComplianceError::AuditError -> AuditLogError
- Add missing 'tags' field to DataMetadata struct

+++ crates/horizon-services/capacity-modeler/src/api/handlers/forecast.rs
@@ -73,7 +73,7 @@ pub async fn forecast_gpu_demand(
 
     // Publish forecast generated event
     state.publish_forecast_event(CapacityMessage::ForecastGenerated {
-        forecast_weeks: request.weeks,
+        forecast_weeks: request.weeks as u32,
         include_confidence: request.include_confidence_intervals,
         timestamp_ms: timestamp_ms(),
     });
@@ -106,8 +106,8 @@ pub async fn backtest_model(
 
     // Publish backtest completed event
     state.publish_forecast_event(CapacityMessage::BacktestCompleted {
-        train_days: backtest_result.train_size,
-        test_days: backtest_result.test_size,
+        train_days: backtest_result.train_size as u32,
+        test_days: backtest_result.test_size as u32,
         mae: backtest_result.metrics.mae,
         mape: backtest_result.metrics.mape,
         timestamp_ms: timestamp_ms(),
